pyproject.toml
[project]
name = "wethepeople"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.128.8",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
]
